requirements.txt
python-dotenv==1.2.2
asyncio==4.0.0
httpx==0.28.1
tonutils==2.0.0
